#34 merges usersLogin test to users test
