+++ lib/users.test.js
@@ -50,3 +50,79 @@ describe('Validates user signup', () => {
     })
   })
 })
+
+jest.mock('../sequelize/db')
+const bcrypt = require('bcrypt')
+const db = require('../sequelize/db')
+
+const genData = async (password) => {
+  return {
+    username: 'testuser',
+    password: await bcrypt.hash(password, 10),
+    email: 'test.user@databases.com'
+  }
+}
+
+db.start = () => {}
+const mockFindOne = jest.fn()
+db.getModels = () => {
+    return {
+        Accounts: {
+            findOne: mockFindOne
+        }
+    }
+}
+
+const logIn = users.logIn
+
+beforeEach(() => {
+    jest.clearAllMocks()
+})
+
+describe('Validates user logIn', () => {
+  it("function should reject if username doesn't exist", () => {
+    const obj = {
+      email: '1234@gmail.com',
+      password: 'password'
+    }
+    return expect(logIn(obj)).resolves.toEqual({
+      errors: [ 'username is a required field' ]
+    })
+  })
+})
+
+describe('Log in', () => {
+  it("should throw error if user is not found", () => {
+    mockFindOne.mockReturnValue(undefined)
+    const obj = {
+      username: 'noexisit',
+      password: '1q2w3e4r',
+      email: 'test.user@databases.com'
+    }
+    return expect(logIn(obj)).rejects.toThrow('User not found')
+  })
+  it("should throw error if password doesn't match", async () => {
+    const dataValues = await genData('abcd1234')
+    mockFindOne.mockReturnValue({
+      dataValues: dataValues
+    })
+    const obj = {
+        username: 'testuser',
+        password: '1q2w3e4r',
+        email: 'test.user@databases.com'
+    }
+    return expect(logIn(obj)).rejects.toThrow('Password does not match')
+  })
+  it("should return user information if it logs in successfully", async () => {
+    const dataValues = await genData('abcd1234')
+    mockFindOne.mockReturnValue({
+      dataValues: dataValues
+    })
+    const obj = {
+        username: 'testuser',
+        password: 'abcd1234',
+        email: 'test.user@databases.com'
+    }
+    return expect(logIn(obj)).resolves.toEqual(dataValues)
+  })
+})
